Added a second button next to logout for navigating to profile page.
When users were logged in and on the home page, they had no way to get to their profile page except to logout and then log back in. This button allows them to skip that and go right to their profile page. I got the buttons to display side by side. Boomski.